Update Sinus approximator sample to SKaiNET 0.2.0

+++ SinusApproximator/composeApp/src/commonMain/kotlin/sk/ai/net/samples/kmp/sinus/approximator/SinusVisualization.kt
@@ -17,7 +17,7 @@ import kotlin.math.sin
 fun SinusVisualization(
     sliderValue: Float,
     actualSinus: Double,
-    approximatedSinus: Double,
+    approximatedSinus: Float,
     modifier: Modifier = Modifier
 ) {
     val primary = MaterialTheme.colorScheme.primary
